T4 Implement multi-threaded broker

+++ core/src/main/java/dev/blechschmidt/quocojms/message/QuotationRequestMessage.java
@@ -5,10 +5,10 @@ import java.io.Serializable;
 import dev.blechschmidt.quocojms.core.ClientInfo;
 
 public class QuotationRequestMessage implements Serializable {
-    public long id;
+    public int id;
     public ClientInfo info;
 
-    public QuotationRequestMessage(long id, ClientInfo info) {
+    public QuotationRequestMessage(int id, ClientInfo info) {
         this.id = id;
         this.info = info;
     }
