+++ core/src/main/java/dev/blechschmidt/quocojms/message/QuotationResponseMessage.java
@@ -5,10 +5,10 @@ import java.io.Serializable;
 import dev.blechschmidt.quocojms.core.Quotation;
 
 public class QuotationResponseMessage implements Serializable {
-    public long id;
+    public int id;
     public Quotation quotation;
 
-    public QuotationResponseMessage(long id, Quotation quotation) {
+    public QuotationResponseMessage(int id, Quotation quotation) {
         this.id = id;
         this.quotation = quotation;
     }
